# lib/api-zod/src/generated/types/expiryScan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExpiryScanStatus } from "./expiryScanStatus";

export interface ExpiryScan {
  id: number;
  sessionId: string;
  pdUserName: string;
  storeLocation: string;
  barcode: string;
  itemNumber?: string | null;
  description?: string | null;
  qty: number;
  expiryDate: Date;
  status: ExpiryScanStatus;
  daysLeft: number;
  scanDate: Date;
  actionRequired?: string | null;
  remarks?: string | null;
  createdAt: Date;
}