# lib/api-zod/src/generated/types/expirySessionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExpirySessionSummary {
  scans: number;
  activeItems: number;
  expiredItems: number;
  totalQty: number;
  urgentItems: number;
  nearExpiryItems: number;
}