# lib/api-zod/src/generated/types/newExpiryScan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewExpiryScan {
  sessionId: string;
  pdUserName: string;
  storeLocation: string;
  barcode: string;
  itemNumber?: string | null;
  description?: string | null;
  /** @minimum 0.01 */
  qty: number;
  expiryDate: Date;
  scanDate: Date;
  remarks?: string | null;
}